refactor: move e2e/support/build-config tests to their own files (Phase 6)
Move the shared test-support helpers + real-client smoke test (mod test), the
loopback e2e suite (mod conformance), and the build-config negative test
(mod feature_config_negative) out of lib.rs into dedicated crate-root files
(test.rs, conformance.rs, feature_config_negative.rs). Test paths (crate::test,
conformance::, feature_config_negative::) and the 2 ignored tests preserved;
CI --exact gate 4 passed. lib.rs now 184 lines. Pure move; no behavior/API change.
